Ajustes no backend e no docker do banco de dados e nos servicos do frontend

+++ Back/src/RS.Parking.Domain/Models/VehicleType.cs
@@ -8,10 +8,14 @@ public class VehicleType
 {
 	#region Properties
 	//[Key]
-	public ulong Id { get; set; }
+	public ushort Id { get; set; }
+
 	public bool Active { get; set; }
+
 	public DateTime DateCreated { get; set; } = DateTime.Now;
+
 	public decimal Cost { get; set; }    //decimal myMoney = 300.5m;
+
 	public string Description { get; set; }
 	#endregion
 }
